unify data transform (#2239)
