+++ dygraph/configs/fcos/_base_/fcos_reader.yml
@@ -1,13 +1,13 @@
 worker_num: 2
 TrainReader:
   sample_transforms:
-  - DecodeOp: {}
-  - RandomFlipOp: {prob: 0.5}
-  - NormalizeImageOp: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
-  - ResizeImage: {target_size: 800, max_size: 1333, interp: 1, use_cv2: true}
-  - PermuteOp: {}
+  - Decode: {}
+  - RandomFlip: {prob: 0.5}
+  - NormalizeImage: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
+  - Resize: {target_size: [800, 1333], keep_ratio: true, interp: 1}
+  - Permute: {}
   batch_transforms:
-  - PadBatchOp: {pad_to_stride: 128}
+  - PadBatch: {pad_to_stride: 128}
   - Gt2FCOSTarget:
       object_sizes_boundary: [64, 128, 256, 512]
       center_sampling_radius: 1.5
@@ -20,23 +20,23 @@ TrainReader:
 
 EvalReader:
   sample_transforms:
-  - DecodeOp: {}
-  - NormalizeImageOp: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
-  - ResizeOp: {interp: 1, target_size: [800, 1333], keep_ratio: True}
-  - PermuteOp: {}
+  - Decode: {}
+  - NormalizeImage: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
+  - Resize: {interp: 1, target_size: [800, 1333], keep_ratio: True}
+  - Permute: {}
   batch_transforms:
-  - PadBatchOp: {pad_to_stride: 128}
+  - PadBatch: {pad_to_stride: 128}
   batch_size: 1
   shuffle: false
 
 
 TestReader:
   sample_transforms:
-  - DecodeOp: {}
-  - NormalizeImageOp: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
-  - ResizeOp: {interp: 1, target_size: [800, 1333], keep_ratio: True}
-  - PermuteOp: {}
+  - Decode: {}
+  - NormalizeImage: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
+  - Resize: {interp: 1, target_size: [800, 1333], keep_ratio: True}
+  - Permute: {}
   batch_transforms:
-  - PadBatchOp: {pad_to_stride: 128}
+  - PadBatch: {pad_to_stride: 128}
   batch_size: 1
   shuffle: false
