+++ dygraph/configs/mask_rcnn/_base_/mask_reader.yml
@@ -1,13 +1,13 @@
 worker_num: 2
 TrainReader:
   sample_transforms:
-  - DecodeOp: {}
-  - RandomResizeOp: {target_size: [[640, 1333], [672, 1333], [704, 1333], [736, 1333], [768, 1333], [800, 1333]], interp: 2, keep_ratio: True}
-  - RandomFlipOp: {prob: 0.5}
-  - NormalizeImageOp: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
-  - PermuteOp: {}
+  - Decode: {}
+  - RandomResize: {target_size: [[640, 1333], [672, 1333], [704, 1333], [736, 1333], [768, 1333], [800, 1333]], interp: 2, keep_ratio: True}
+  - RandomFlip: {prob: 0.5}
+  - NormalizeImage: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
+  - Permute: {}
   batch_transforms:
-  - PadBatchOp: {pad_to_stride: -1., pad_gt: true}
+  - PadBatch: {pad_to_stride: -1., pad_gt: true}
   batch_size: 1
   shuffle: true
   drop_last: true
@@ -15,12 +15,12 @@ TrainReader:
 
 EvalReader:
   sample_transforms:
-  - DecodeOp: {}
-  - ResizeOp: {interp: 2, target_size: [800, 1333], keep_ratio: True}
-  - NormalizeImageOp: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
-  - PermuteOp: {}
+  - Decode: {}
+  - Resize: {interp: 2, target_size: [800, 1333], keep_ratio: True}
+  - NormalizeImage: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
+  - Permute: {}
   batch_transforms:
-  - PadBatchOp: {pad_to_stride: -1., pad_gt: false}
+  - PadBatch: {pad_to_stride: -1., pad_gt: false}
   batch_size: 1
   shuffle: false
   drop_last: false
@@ -29,12 +29,12 @@ EvalReader:
 
 TestReader:
   sample_transforms:
-  - DecodeOp: {}
-  - ResizeOp: {interp: 2, target_size: [800, 1333], keep_ratio: True}
-  - NormalizeImageOp: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
-  - PermuteOp: {}
+  - Decode: {}
+  - Resize: {interp: 2, target_size: [800, 1333], keep_ratio: True}
+  - NormalizeImage: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
+  - Permute: {}
   batch_transforms:
-  - PadBatchOp: {pad_to_stride: -1., pad_gt: false}
+  - PadBatch: {pad_to_stride: -1., pad_gt: false}
   batch_size: 1
   shuffle: false
   drop_last: false
