+++ dygraph/configs/solov2/_base_/solov2_reader.yml
@@ -1,15 +1,15 @@
 worker_num: 2
 TrainReader:
   sample_transforms:
-  - DecodeOp: {}
+  - Decode: {}
   - Poly2Mask: {}
-  - ResizeOp: {interp: 1, target_size: [800, 1333], keep_ratio: True}
-  - RandomFlipOp: {}
-  - NormalizeImageOp: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
-  - PermuteOp: {}
+  - Resize: {interp: 1, target_size: [800, 1333], keep_ratio: True}
+  - RandomFlip: {}
+  - NormalizeImage: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
+  - Permute: {}
   batch_transforms:
-  - PadBatchOp: {pad_to_stride: 32}
-  - Gt2Solov2TargetOp: {num_grids: [40, 36, 24, 16, 12],
+  - PadBatch: {pad_to_stride: 32}
+  - Gt2Solov2Target: {num_grids: [40, 36, 24, 16, 12],
                         scale_ranges: [[1, 96], [48, 192], [96, 384], [192, 768], [384, 2048]],
                         coord_sigma: 0.2}
   batch_size: 2
@@ -19,12 +19,12 @@ TrainReader:
 
 EvalReader:
   sample_transforms:
-  - DecodeOp: {}
-  - NormalizeImageOp: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
-  - ResizeOp: {interp: 1, target_size: [800, 1333], keep_ratio: True}
-  - PermuteOp: {}
+  - Decode: {}
+  - NormalizeImage: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
+  - Resize: {interp: 1, target_size: [800, 1333], keep_ratio: True}
+  - Permute: {}
   batch_transforms:
-  - PadBatchOp: {pad_to_stride: 32}
+  - PadBatch: {pad_to_stride: 32}
   batch_size: 1
   shuffle: false
   drop_last: false
@@ -33,12 +33,12 @@ EvalReader:
 
 TestReader:
   sample_transforms:
-  - DecodeOp: {}
-  - NormalizeImageOp: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
-  - ResizeOp: {interp: 1, target_size: [800, 1333], keep_ratio: True}
-  - PermuteOp: {}
+  - Decode: {}
+  - NormalizeImage: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
+  - Resize: {interp: 1, target_size: [800, 1333], keep_ratio: True}
+  - Permute: {}
   batch_transforms:
-  - PadBatchOp: {pad_to_stride: 32}
+  - PadBatch: {pad_to_stride: 32}
   batch_size: 1
   shuffle: false
   drop_last: false
