+++ dygraph/configs/ssd/_base_/ssd_mobilenet_reader.yml
@@ -3,17 +3,17 @@ TrainReader:
   inputs_def:
     num_max_boxes: 90
   sample_transforms:
-    - DecodeOp: {}
-    - RandomDistortOp: {brightness: [0.5, 1.125, 0.875], random_apply: False}
-    - RandomExpandOp: {fill_value: [127.5, 127.5, 127.5]}
-    - RandomCropOp: {allow_no_crop: Fasle}
-    - RandomFlipOp: {}
-    - ResizeOp: {target_size: [300, 300], keep_ratio: False, interp: 1}
-    - NormalizeBoxOp: {}
-    - PadBoxOp: {num_max_boxes: 90}
+    - Decode: {}
+    - RandomDistort: {brightness: [0.5, 1.125, 0.875], random_apply: False}
+    - RandomExpand: {fill_value: [127.5, 127.5, 127.5]}
+    - RandomCrop: {allow_no_crop: Fasle}
+    - RandomFlip: {}
+    - Resize: {target_size: [300, 300], keep_ratio: False, interp: 1}
+    - NormalizeBox: {}
+    - PadBox: {num_max_boxes: 90}
   batch_transforms:
-    - NormalizeImageOp: {mean: [127.5, 127.5, 127.5], std: [127.502231, 127.502231, 127.502231], is_scale: false}
-    - PermuteOp: {}
+    - NormalizeImage: {mean: [127.5, 127.5, 127.5], std: [127.502231, 127.502231, 127.502231], is_scale: false}
+    - Permute: {}
   batch_size: 32
   shuffle: true
   drop_last: true
@@ -21,10 +21,10 @@ TrainReader:
 
 EvalReader:
   sample_transforms:
-    - DecodeOp: {}
-    - ResizeOp: {target_size: [300, 300], keep_ratio: False, interp: 1}
-    - NormalizeImageOp: {mean: [127.5, 127.5, 127.5], std: [127.502231, 127.502231, 127.502231], is_scale: false}
-    - PermuteOp: {}
+    - Decode: {}
+    - Resize: {target_size: [300, 300], keep_ratio: False, interp: 1}
+    - NormalizeImage: {mean: [127.5, 127.5, 127.5], std: [127.502231, 127.502231, 127.502231], is_scale: false}
+    - Permute: {}
   batch_size: 1
   drop_empty: false
 
@@ -33,8 +33,8 @@ TestReader:
   inputs_def:
     image_shape: [3, 300, 300]
   sample_transforms:
-    - DecodeOp: {}
-    - ResizeOp: {target_size: [300, 300], keep_ratio: False, interp: 1}
-    - NormalizeImageOp: {mean: [127.5, 127.5, 127.5], std: [127.502231, 127.502231, 127.502231], is_scale: false}
-    - PermuteOp: {}
+    - Decode: {}
+    - Resize: {target_size: [300, 300], keep_ratio: False, interp: 1}
+    - NormalizeImage: {mean: [127.5, 127.5, 127.5], std: [127.502231, 127.502231, 127.502231], is_scale: false}
+    - Permute: {}
   batch_size: 1
