+++ dygraph/configs/ssd/_base_/ssd_reader.yml
@@ -4,18 +4,18 @@ TrainReader:
     num_max_boxes: 90
 
   sample_transforms:
-    - DecodeOp: {}
-    - RandomDistortOp: {brightness: [0.5, 1.125, 0.875], random_apply: False}
-    - RandomExpandOp: {fill_value: [104., 117., 123.]}
-    - RandomCropOp: {allow_no_crop: true}
-    - RandomFlipOp: {}
-    - ResizeOp: {target_size: [300, 300], keep_ratio: False, interp: 1}
-    - NormalizeBoxOp: {}
-    - PadBoxOp: {num_max_boxes: 90}
+    - Decode: {}
+    - RandomDistort: {brightness: [0.5, 1.125, 0.875], random_apply: False}
+    - RandomExpand: {fill_value: [104., 117., 123.]}
+    - RandomCrop: {allow_no_crop: true}
+    - RandomFlip: {}
+    - Resize: {target_size: [300, 300], keep_ratio: False, interp: 1}
+    - NormalizeBox: {}
+    - PadBox: {num_max_boxes: 90}
 
   batch_transforms:
-    - NormalizeImageOp: {mean: [104., 117., 123.], std: [1., 1., 1.], is_scale: false}
-    - PermuteOp: {}
+    - NormalizeImage: {mean: [104., 117., 123.], std: [1., 1., 1.], is_scale: false}
+    - Permute: {}
 
   batch_size: 8
   shuffle: true
@@ -24,10 +24,10 @@ TrainReader:
 
 EvalReader:
   sample_transforms:
-    - DecodeOp: {}
-    - ResizeOp: {target_size: [300, 300], keep_ratio: False, interp: 1}
-    - NormalizeImageOp: {mean: [104., 117., 123.], std: [1., 1., 1.], is_scale: false}
-    - PermuteOp: {}
+    - Decode: {}
+    - Resize: {target_size: [300, 300], keep_ratio: False, interp: 1}
+    - NormalizeImage: {mean: [104., 117., 123.], std: [1., 1., 1.], is_scale: false}
+    - Permute: {}
   batch_size: 1
   drop_empty: false
 
@@ -35,8 +35,8 @@ TestReader:
   inputs_def:
     image_shape: [3, 300, 300]
   sample_transforms:
-    - DecodeOp: {}
-    - ResizeOp: {target_size: [300, 300], keep_ratio: False, interp: 1}
-    - NormalizeImageOp: {mean: [104., 117., 123.], std: [1., 1., 1.], is_scale: false}
-    - PermuteOp: {}
+    - Decode: {}
+    - Resize: {target_size: [300, 300], keep_ratio: False, interp: 1}
+    - NormalizeImage: {mean: [104., 117., 123.], std: [1., 1., 1.], is_scale: false}
+    - Permute: {}
   batch_size: 1
