+++ dygraph/configs/ssd/_base_/ssdlite320_reader.yml
@@ -3,17 +3,17 @@ TrainReader:
   inputs_def:
     num_max_boxes: 90
   sample_transforms:
-    - DecodeOp: {}
-    - RandomDistortOp: {brightness: [0.5, 1.125, 0.875], random_apply: False}
-    - RandomExpandOp: {fill_value: [123.675, 116.28, 103.53]}
-    - RandomCropOp: {allow_no_crop: Fasle}
-    - RandomFlipOp: {}
-    - ResizeOp: {target_size: [320, 320], keep_ratio: False, interp: 1}
-    - NormalizeBoxOp: {}
-    - PadBoxOp: {num_max_boxes: 90}
+    - Decode: {}
+    - RandomDistort: {brightness: [0.5, 1.125, 0.875], random_apply: False}
+    - RandomExpand: {fill_value: [123.675, 116.28, 103.53]}
+    - RandomCrop: {allow_no_crop: Fasle}
+    - RandomFlip: {}
+    - Resize: {target_size: [320, 320], keep_ratio: False, interp: 1}
+    - NormalizeBox: {}
+    - PadBox: {num_max_boxes: 90}
   batch_transforms:
-    - NormalizeImageOp: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: true}
-    - PermuteOp: {}
+    - NormalizeImage: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: true}
+    - Permute: {}
   batch_size: 64
   shuffle: true
   drop_last: true
@@ -21,10 +21,10 @@ TrainReader:
 
 EvalReader:
   sample_transforms:
-    - DecodeOp: {}
-    - ResizeOp: {target_size: [320, 320], keep_ratio: False, interp: 1}
-    - NormalizeImageOp: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: true}
-    - PermuteOp: {}
+    - Decode: {}
+    - Resize: {target_size: [320, 320], keep_ratio: False, interp: 1}
+    - NormalizeImage: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: true}
+    - Permute: {}
   batch_size: 1
   drop_empty: false
 
@@ -33,8 +33,8 @@ TestReader:
   inputs_def:
     image_shape: [3, 320, 320]
   sample_transforms:
-    - DecodeOp: {}
-    - ResizeOp: {target_size: [320, 320], keep_ratio: False, interp: 1}
-    - NormalizeImageOp: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: true}
-    - PermuteOp: {}
+    - Decode: {}
+    - Resize: {target_size: [320, 320], keep_ratio: False, interp: 1}
+    - NormalizeImage: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: true}
+    - Permute: {}
   batch_size: 1
