+++ dygraph/configs/ttfnet/_base_/ttfnet_reader.yml
@@ -1,34 +1,34 @@
 worker_num: 2
 TrainReader:
   sample_transforms:
-  - DecodeOp: {}
-  - RandomFlipOp: {prob: 0.5}
-  - ResizeOp: {interp: 1, target_size: [512, 512], keep_ratio: False}
-  - NormalizeImageOp: {mean: [123.675, 116.28, 103.53], std: [58.395, 57.12, 57.375], is_scale: false}
-  - PermuteOp: {}
+  - Decode: {}
+  - RandomFlip: {prob: 0.5}
+  - Resize: {interp: 1, target_size: [512, 512], keep_ratio: False}
+  - NormalizeImage: {mean: [123.675, 116.28, 103.53], std: [58.395, 57.12, 57.375], is_scale: false}
+  - Permute: {}
   batch_transforms:
-  - Gt2TTFTargetOp: {down_ratio: 4}
-  - PadBatchOp: {pad_to_stride: 32, pad_gt: true}
+  - Gt2TTFTarget: {down_ratio: 4}
+  - PadBatch: {pad_to_stride: 32, pad_gt: true}
   batch_size: 12
   shuffle: true
   drop_last: true
 
 EvalReader:
   sample_transforms:
-  - DecodeOp: {}
-  - ResizeOp: {interp: 1, target_size: [512, 512], keep_ratio: False}
-  - NormalizeImageOp: {is_scale: false, mean: [123.675, 116.28, 103.53], std: [58.395, 57.12, 57.375]}
-  - PermuteOp: {}
+  - Decode: {}
+  - Resize: {interp: 1, target_size: [512, 512], keep_ratio: False}
+  - NormalizeImage: {is_scale: false, mean: [123.675, 116.28, 103.53], std: [58.395, 57.12, 57.375]}
+  - Permute: {}
   batch_size: 1
   drop_last: false
   drop_empty: false
 
 TestReader:
   sample_transforms:
-  - DecodeOp: {}
-  - ResizeOp: {interp: 1, target_size: [512, 512], keep_ratio: False}
-  - NormalizeImageOp: {is_scale: false, mean: [123.675, 116.28, 103.53], std: [58.395, 57.12, 57.375]}
-  - PermuteOp: {}
+  - Decode: {}
+  - Resize: {interp: 1, target_size: [512, 512], keep_ratio: False}
+  - NormalizeImage: {is_scale: false, mean: [123.675, 116.28, 103.53], std: [58.395, 57.12, 57.375]}
+  - Permute: {}
   batch_size: 1
   drop_last: false
   drop_empty: false
