+++ dygraph/configs/yolov3/_base_/yolov3_reader.yml
@@ -3,20 +3,20 @@ TrainReader:
   inputs_def:
     num_max_boxes: 50
   sample_transforms:
-    - DecodeOp: {}
-    - MixupOp: {alpha: 1.5, beta: 1.5}
-    - RandomDistortOp: {}
-    - RandomExpandOp: {fill_value: [123.675, 116.28, 103.53]}
-    - RandomCropOp: {}
-    - RandomFlipOp: {}
+    - Decode: {}
+    - Mixup: {alpha: 1.5, beta: 1.5}
+    - RandomDistort: {}
+    - RandomExpand: {fill_value: [123.675, 116.28, 103.53]}
+    - RandomCrop: {}
+    - RandomFlip: {}
   batch_transforms:
-    - BatchRandomResizeOp: {target_size: [320, 352, 384, 416, 448, 480, 512, 544, 576, 608], random_size: True, random_interp: True, keep_ratio: False}
-    - NormalizeBoxOp: {}
-    - PadBoxOp: {num_max_boxes: 50}
-    - BboxXYXY2XYWHOp: {}
-    - NormalizeImageOp: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: True}
-    - PermuteOp: {}
-    - Gt2YoloTargetOp: {anchor_masks: [[6, 7, 8], [3, 4, 5], [0, 1, 2]], anchors: [[10, 13], [16, 30], [33, 23], [30, 61], [62, 45], [59, 119], [116, 90], [156, 198], [373, 326]], downsample_ratios: [32, 16, 8]}
+    - BatchRandomResize: {target_size: [320, 352, 384, 416, 448, 480, 512, 544, 576, 608], random_size: True, random_interp: True, keep_ratio: False}
+    - NormalizeBox: {}
+    - PadBox: {num_max_boxes: 50}
+    - BboxXYXY2XYWH: {}
+    - NormalizeImage: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: True}
+    - Permute: {}
+    - Gt2YoloTarget: {anchor_masks: [[6, 7, 8], [3, 4, 5], [0, 1, 2]], anchors: [[10, 13], [16, 30], [33, 23], [30, 61], [62, 45], [59, 119], [116, 90], [156, 198], [373, 326]], downsample_ratios: [32, 16, 8]}
   batch_size: 8
   shuffle: true
   drop_last: true
@@ -27,10 +27,10 @@ EvalReader:
   inputs_def:
     num_max_boxes: 50
   sample_transforms:
-    - DecodeOp: {}
-    - ResizeOp: {target_size: [608, 608], keep_ratio: False, interp: 2}
-    - NormalizeImageOp: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: True}
-    - PermuteOp: {}
+    - Decode: {}
+    - Resize: {target_size: [608, 608], keep_ratio: False, interp: 2}
+    - NormalizeImage: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: True}
+    - Permute: {}
   batch_size: 1
   drop_empty: false
 
@@ -38,8 +38,8 @@ TestReader:
   inputs_def:
     image_shape: [3, 608, 608]
   sample_transforms:
-    - DecodeOp: {}
-    - ResizeOp: {target_size: [608, 608], keep_ratio: False, interp: 2}
-    - NormalizeImageOp: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: True}
-    - PermuteOp: {}
+    - Decode: {}
+    - Resize: {target_size: [608, 608], keep_ratio: False, interp: 2}
+    - NormalizeImage: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: True}
+    - Permute: {}
   batch_size: 1
